cast predictions,labels to Double

+++ src/main/scala/com/sgcharts/badrenter/ModelTesting.scala
@@ -63,8 +63,8 @@ object ModelTesting extends Log4jLogging {
     val transformed: DataFrame = model.transform(data)
     transformed.printSchema()
     val res = transformed.rdd.map { row =>
-        val pred = row.getAs[Double]("prediction")
-        val label = row.getAs[Double]("label")
+        val pred = row.getAs[Int]("prediction").toDouble
+        val label = row.getAs[Int]("label").toDouble
         (pred, label)
     }
     val metrics = new RegressionMetrics(res)
